add kyc framework bone, and refactor expression
