+++ services/kyc/Cargo.toml
@@ -8,3 +8,9 @@ edition = "2018"
 
 [dependencies]
 derive_more = "0.99"
+muta = { git = "https://github.com/nervosnetwork/muta", tag = "v0.1.2-beta2"}
+protocol = { git = "https://github.com/nervosnetwork/muta", tag = "v0.1.2-beta2", package = "muta-protocol" }
+binding-macro = { git = "https://github.com/nervosnetwork/muta", tag = "v0.1.2-beta2" }
+
+[dev-dependencies]
+framework = { git = "https://github.com/nervosnetwork/muta", tag = "v0.1.2-beta2"}
